Fixed a typing issue. Makes me miss Python and Lisp... :'(

+++ src/com/a9development/ciphertestharness/RijndaelCipherTest.java
@@ -29,7 +29,7 @@ public class RijndaelCipherTest {
 	@Test
 	public void testRijndaelDecrypt() throws Exception {
 		RijndaelCipher rijndaelTester1 = new RijndaelCipher(K1);
-		int[] result = rijndaelTester1.decrypt(CT1);
+		byte[] result = rijndaelTester1.decrypt(CT1);
 		String resultString = "";
 		for (int i = 0; i < 16; i++) {
 			resultString += Integer.toString((result[i] & 0xff) + 0x100, 16).substring(1);
